+++ server/src/services/util.ts
@@ -1,4 +1,10 @@
-import { Document } from '../types/apiInterface';
+import {
+  Document,
+  DocumentsCreate,
+  DocumentsUpdate,
+  keyofDocumentsCreate,
+  keyofDocumentsUpdate,
+} from '../types/apiInterface';
 
 export function getObjectKey(arg: object): string[] {
   return Object.entries(arg)
@@ -12,8 +18,26 @@ export function getObjectValue(arg: object): string[] {
     .map(([, value]) => `\'${value}\'`);
 }
 
-export function getDocumentKeyValue(arg: Document, stringTypeList: String[]): String[] {
+export function getDocumentsCreateObj(param: DocumentsCreate): object {
+  const result = {};
+  Object.entries(keyofDocumentsCreate).forEach(([key]) => (result[key] = param[key]));
+  return result;
+}
+
+export function getDocumentsUpdateObj(param: DocumentsUpdate): object {
+  const result = {};
+  Object.entries(keyofDocumentsUpdate).forEach(([key]) => (result[key] = param[key]));
+  return result;
+}
+
+export function getDocumentKeyValue(arg: Document, stringTypeList: String[], append: String = undefined): String[] {
   return Object.entries(arg)
     .filter(([, value]) => value !== undefined && value !== null)
-    .map(([key, value]) => `${key}=${!stringTypeList.includes(key) ? `'${value}'` : value}`);
+    .map(([key, value]) => {
+      let _key = key;
+      if (append || append.length > 0) {
+        _key = `${append}${key}`;
+      }
+      return `${_key}=${!stringTypeList.includes(key) ? `'${value}'` : value}`;
+    });
 }
